docs: Update CONTRIBUTING
Tighten up some of the explanations and update obsolete parts of the
CONTRIBUTING document.

This doesn't change any of our policies or procedures. But some advice
was out of date or no longer applicable (for example, now that we use
clang-format, we do not need to detail all the formatting rules, they
are enforced automatically).

Signed-off-by: Larry Gritz <lg@larrygritz.com>