ci: Switch OIIO dev-2.3 branch to using v2.3.21.0 tag
OIIO's 2.3 family is obsolete and no longer updating, so we removed
the branch marker. (Branch markers should always indicate places where
development may happen.)

But we sill reference that branch namd in a couple places for OSL's CI,
so just replace it with the immutable tag of the last release in the 2.3
family.

Signed-off-by: Larry Gritz <lg@larrygritz.com>