build(deps): Raise minimum OpenImageIO version to 2.4
This is just for future-looking OSL 1.13 releases, it won't be
backported to the 1.12 release branch.

Remember that by the time we release OSL 1.13 as a supported release
family, OIIO will also have done its annual release, of 2.5. So our
minimum of OIIO 2.4 means that we will be supporting the current
release, as well as the prior year's release (and, of course, we
always stay compatible with OIIO's trunk as it evolves). So we're
really keeping compatibility with 3 major releases, which seems like
plenty to juggle.

Signed-off-by: Larry Gritz <lg@larrygritz.com>