LLVM opaque pointers support (#1728)
* LLVM opaque pointers support

New LLVM versions require opaque pointers, that is you can no longer tell
from a pointer value which data type it points to.

* Remove deprecated GEP and op_load methods without type, and refactor
  code to always pass the type.
* Add element type argument to op_scatter and op_gather.
* Eliminate pointer type comparison which no longer work with
  opaque pointers.
* Keep interactive params marked as uniform in batched analysis, so
  we can deduce uniform/varying correctly from just the symbol
* Enable opaque pointers on LLVM 15, except when using OptiX

OptiX opaque pointers were not tested yet and left disabled. It is
likely broken due to the sneaky conversion in
`BackendLLVM::llvm_store_value` between `int64` and `char*`, but we
can no longer compare the pointer type. This function is called from
many places and it's unclear to me when this happens.

According to a4e779c4 this code is only temporary, so this may resolve
itself as the ustring refactoring gets completed.

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* Fix debug_uninit and pointcloud batched testsß

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* Variable renaming and comment for clarity

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* Add/restore debug names for groupdata access

Make batched backend more consistent with the regular backend.

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* Rename scalar_type to component_type

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* Alternative interactive param fix based on feedback from Alex

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* OptiX fixes

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

* Use opaque pointers only for LLVM 16+

To avoid potentially breaking LLVM 15 with OptiX.

Signed-off-by: Brecht Van Lommel <brecht@blender.org>

---------

Signed-off-by: Brecht Van Lommel <brecht@blender.org>