fix: SymOverrideInfo bitfields should be the same type (#1745)
Stephen Friedman alerted me to the fact that on Windows, bit fields
within a struct will only be combined if they are the same type.  We
were a bit sloppy in the defintion of SymOverrideInfo, and that was
causing the struct to end up a different size in Windows.  I think the
solution is that all those bit fields should be `unsigned int`.

---------

Signed-off-by: Larry Gritz <lg@larrygritz.com>