ci: Increase timeout for slow MacOS-13 runner
We didn't change anything on our end, but in the last few days, I
believe the GHA MacOS-13 runners have slowed done somehow, and so now
a few test are failing only because they don't complete within the
ctest timeout period for "small tests". So this just bumps up the
limit on this runner platform to allow the tests to pass.

Signed-off-by: Larry Gritz <lg@larrygritz.com>