build: print commit hash of locally built dependencies
Especially for our "bleeding edge" test that builds from the current
master of several dependencies, sometimes our tests fail not because
of anything we've done, but because of a change in one of the
dependencies.  To help track down exactly which change broke the
dependency, make sure our CI build logs print the specific commit hash
of those packages that we built.

Signed-off-by: Larry Gritz <lg@larrygritz.com>